Issue #5: receive long from transmitter as value exceeds int8

+++ WaterLevelReceiver-ESP8266/RF433Receiver.ino
@@ -10,9 +10,9 @@ void setup_receiver() {
   mySwitch.enableReceive(D4);
 }
 
-int receive() {
+long receive() {
   if (mySwitch.available()) {
-    int level = mySwitch.getReceivedValue();
+    long level = mySwitch.getReceivedValue();
     mySwitch.resetAvailable();
     return level;
   } else {
